+++ bin/check-cloudwatch-metric.rb
@@ -27,11 +27,17 @@
 #   for details.
 #
 
-require 'sensu-plugins-aws/cloudwatch-common'
+require 'sensu-plugins-aws'
 require 'sensu-plugin/check/cli'
 require 'aws-sdk'
 
 class CloudWatchMetricCheck < Sensu::Plugin::Check::CLI
+  option :aws_region,
+         short: '-r AWS_REGION',
+         long: '--aws-region REGION',
+         description: 'AWS Region (defaults to us-east-1).',
+         default: 'us-east-1'
+
   option :namespace,
          description: 'CloudWatch namespace for metric',
          short: '-n NAME',