+++ lib/sensu-plugins-aws/cloudwatch-common.rb
@@ -1,6 +1,7 @@
-require 'aws-sdk'
 
 module CloudwatchCommon
+  include Common
+
   def client
     Aws::CloudWatch::Client.new
   end